+++ apollo-federation-types/CHANGELOG.md
@@ -4,7 +4,11 @@ Not every version is listed here because versions before 0.14.0 did not have a c
 
 ## 0.16.1
 
+- Update `apollo-federation` dependency to v2.11.0-abstract.1
+
 - Update `apollo-federation` dependency to v2.8.2 (from v2.8.0)
+
+- Update `apollo-federation` dependency to v2.8.0-preview.0 (from v2.7.0)
 
 ## 0.16.0